same

+++ core/class/philipsHue.class.php
@@ -34,10 +34,12 @@ class philipsHue extends eqLogic {
 		$return['state'] = 'nok';
 		$pid_file = jeedom::getTmpFolder('philipsHue') . '/deamon.pid';
 		if (file_exists($pid_file)) {
-			if (trim(file_get_contents($pid_file)) === '') {
-				shell_exec(system::getCmdSudo() . 'rm -rf ' . $pid_file . ' 2>&1 > /dev/null');
-			}else if (@posix_getsid((int) trim(file_get_contents($pid_file)))) {
+			if (trim(file_get_contents($pid_file)) != '' && @posix_getsid((int)trim(file_get_contents($pid_file)))) {
 				$return['state'] = 'ok';
+			} else {
+				if (trim(file_get_contents($pid_file)) != '') {
+				   shell_exec(system::getCmdSudo() . 'rm -rf ' . $pid_file . ' 2>&1 > /dev/null');
+				}
 			}
 		}
 		$return['launchable'] = 'ok';
